Progress on refactor
