+++ index.js
@@ -1,64 +1,30 @@
-const binding = require('./lib/binding')
-const Connection = require('./lib/connection')
 const EventEmitter = require('events')
+const net = require('net')
 const dns = require('dns')
-const set = require('unordered-set')
+const { Duplex } = require('streamx')
+const timeout = require('timeout-refresh')
 const b4a = require('b4a')
-const queueTick = require('queue-tick')
-
-const EMPTY = b4a.alloc(0)
-const IPv4Pattern = /^((?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])[.]){3}(?:[0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$/
+const binding = require('./lib/binding')
+const UTPConnection = require('./lib/utp-connection')
+const UTPSocket = require('./lib/utp-socket')
 
 const Socket = module.exports = class Socket extends EventEmitter {
   constructor (opts) {
     super()
 
     this.connections = []
-
-    this._sending = []
-    this._sent = []
-    this._offset = 0
-    this._buffer = b4a.allocUnsafe(2 * 65536)
-    this._handle = b4a.alloc(binding.sizeof_utp_napi_t)
-    this._nextConnection = b4a.alloc(binding.sizeof_utp_napi_connection_t)
-    this._address = null
-    this._inited = false
-    this._refed = true
-    this._closing = false
-    this._closed = false
-    this._allowHalfOpen = !opts || opts.allowHalfOpen !== false
-    this._acceptConnections = new Uint32Array(this._handle.buffer, this._handle.byteOffset + binding.offsetof_utp_napi_t_accept_connections, 1)
     this.maxConnections = 0
+
+    this._address = null
+    this._socket = new UTPSocket()
+    this._allowHalfOpen = !opts || opts.allowHalfOpen !== false
+
+    this._socket
+      .on('connection', this._onconnection.bind(this))
   }
 
-  _init () {
-    this._inited = true
-
-    binding.utp_napi_init(this._handle, this,
-      this._nextConnection,
-      this._buffer,
-      this._onmessage,
-      this._onsend,
-      this._onconnection,
-      this._onclose,
-      this._realloc
-    )
-
-    if (!this._refed) this.unref()
-  }
-
-  firewall (yes) {
-    this._acceptConnections[0] = yes ? 0 : 1
-  }
-
-  ref () {
-    if (this._inited) binding.utp_napi_ref(this._handle)
-    this._refed = true
-  }
-
-  unref () {
-    if (this._inited) binding.utp_napi_unref(this._handle)
-    this._refed = false
+  firewall (enable) {
+    this._socket.firewall(enable)
   }
 
   address () {
@@ -66,98 +32,67 @@ const Socket = module.exports = class Socket extends EventEmitter {
     return {
       address: this._address,
       family: 'IPv4',
-      port: binding.utp_napi_local_port(this._handle)
+      port: binding.utp_napi_local_port(this._socket._handle)
     }
   }
 
   getRecvBufferSize () {
     if (!this._inited) throw new Error('getRecvBufferSize EBADF')
     if (this._closing) return 0
-    return binding.utp_napi_recv_buffer(this._handle, 0)
+    return binding.utp_napi_recv_buffer(this._socket._handle, 0)
   }
 
   setRecvBufferSize (n) {
     if (!this._inited) throw new Error('setRecvBufferSize EBADF')
     if (this._closing) return 0
-    return binding.utp_napi_recv_buffer(this._handle, n)
+    return binding.utp_napi_recv_buffer(this._socket._handle, n)
   }
 
   getSendBufferSize () {
     if (!this._inited) throw new Error('getSendBufferSize EBADF')
     if (this._closing) return 0
-    return binding.utp_napi_send_buffer(this._handle, 0)
+    return binding.utp_napi_send_buffer(this._socket._handle, 0)
   }
 
   setSendBufferSize (n) {
     if (!this._inited) throw new Error('setSendBufferSize EBADF')
     if (this._closing) return 0
-    return binding.utp_napi_send_buffer(this._handle, n)
+    return binding.utp_napi_send_buffer(this._socket._handle, n)
   }
 
   setTTL (ttl) {
     if (!this._inited) throw new Error('setTTL EBADF')
     if (this._closing) return
-    binding.utp_napi_set_ttl(this._handle, ttl)
-  }
-
-  send (buf, offset, len, port, host, cb) {
-    if (!cb) cb = noop
-    if (!isIP(host)) return this._resolveAndSend(buf, offset, len, port, host, cb)
-    if (this._closing) return queueTick(() => cb(new Error('Socket is closed')))
-    if (!this._address) this.bind(0)
-
-    var send = this._sent.pop()
-    if (!send) {
-      send = new SendRequest()
-      binding.utp_napi_send_request_init(send._handle, send)
-    }
-
-    send._index = this._sending.push(send) - 1
-    send._buffer = buf
-    send._callback = cb
-
-    binding.utp_napi_send(this._handle, send._handle, send._buffer, offset, len, port, host)
-  }
-
-  _resolveAndSend (buf, offset, len, port, host, cb) {
-    const self = this
-
-    dns.lookup(host, { family: 4 }, onlookup)
-
-    function onlookup (err, ip) {
-      if (err) return cb(err)
-      if (!ip) return cb(new Error('Could not resolve ' + host))
-      self.send(buf, offset, len, port, ip, cb)
-    }
-  }
-
-  close (cb) {
-    if (this._closed) return queueTick(() => cb && cb())
-    if (cb) this.once('close', cb)
-    if (this._closing) return
-    this._closing = true
-    this._closeMaybe()
-  }
-
-  _closeMaybe () {
-    if (this._closing && !this.connections.length && !this._sending.length && this._inited && !this._closed) {
-      this._closed = true
-      binding.utp_napi_close(this._handle)
-    }
+    binding.utp_napi_set_ttl(this._socket._handle, ttl)
   }
 
   connect (port, ip) {
-    if (!this._inited) this.bind()
+    if (!this._socket.bound) this.bind()
     if (!ip) ip = '127.0.0.1'
-    const conn = new Connection(this, port, ip, null, this._allowHalfOpen)
-    if (!isIP(ip)) conn._resolveAndConnect(port, ip)
-    else conn._connect(port, ip || '127.0.0.1')
-    return conn
+
+    const connection = new Connection(
+      this,
+      null,
+      port,
+      ip,
+      this._allowHalfOpen
+    )
+
+    if (!net.isIPv4(ip)) connection._resolveAndConnect(port, ip)
+    else connection._connect(port, ip)
+
+    return connection
   }
 
   listen (port, ip, onlistening) {
-    if (!this._address) this.bind(port, ip, onlistening)
+    if (!this._socket.bound) this.bind(port, ip, onlistening)
     this.firewall(false)
+  }
+
+  send (buf, offset, len, port, host, cb) {
+    if (!this._socket.bound) this._socket.bind()
+    if (!net.isIPv4(host)) return this._resolveAndSend(buf, offset, len, port, host, cb)
+    if (this._closing) return cb(new Error('Socket is closed'))
   }
 
   bind (port, ip, onlistening) {
@@ -166,8 +101,7 @@ const Socket = module.exports = class Socket extends EventEmitter {
     if (!port) port = 0
     if (!ip) ip = '0.0.0.0'
 
-    if (!this._inited) this._init()
-    if (this._closing) return
+    if (this._socket.closing) return
 
     if (this._address) {
       this.emit('error', new Error('Socket already bound'))
@@ -175,76 +109,57 @@ const Socket = module.exports = class Socket extends EventEmitter {
     }
 
     if (onlistening) this.once('listening', onlistening)
-    if (!isIP(ip)) return this._resolveAndBind(port, ip)
+    if (!net.isIPv4(ip)) return this._resolveAndBind(port, ip)
 
     this._address = ip
 
-    try {
-      binding.utp_napi_bind(this._handle, port, ip)
-    } catch (err) {
-      this._address = null
-      queueTick(() => this.emit('error', err))
-      return
-    }
-
-    queueTick(() => this.emit('listening'))
-  }
-
-  _resolveAndBind (port, host) {
-    const self = this
-
-    dns.lookup(host, { family: 4 }, function (err, ip) {
-      if (err) return self.emit('error', err)
-      self.bind(port, ip)
+    this._socket.bind(port, ip, (err) => {
+      if (err) {
+        this._address = null
+        this.emit('error', err)
+      } else {
+        this.emit('listening')
+      }
     })
   }
 
-  _realloc () {
-    this._buffer = b4a.allocUnsafe(this._buffer.length)
-    this._offset = 0
-    return this._buffer
+  close (cb) {
+    if (this._socket.closed) return cb && cb(null)
+    if (cb) this.once('close', cb)
+    if (this._socket.closing) return
+
+    this._socket.close((err) => {
+      if (err) this.emit('error', err)
+    })
   }
 
-  _onmessage (size, port, address) {
-    if (size < 0) {
-      this.emit('error', new Error('Read failed (status: ' + size + ')'))
-      return EMPTY
-    }
-
-    const message = this._buffer.subarray(this._offset, this._offset += size)
-    this.emit('message', message, { address, family: 'IPv4', port })
-
-    if (this._buffer.length - this._offset <= 65536) {
-      this._buffer = b4a.allocUnsafe(this._buffer.length)
-      this._offset = 0
-      return this._buffer
-    }
-
-    return EMPTY
+  _closeMaybe () {
+    // if (this._closing && !this.connections.length && !this._sending.length && this._inited && !this._closed) {
+    //   this._closed = true
+    //   binding.utp_napi_close(this._handle)
+    // }
   }
 
-  _onsend (send, status) {
-    const cb = send._callback
-
-    send._callback = send._buffer = null
-    set.remove(this._sending, send)
-    this._sent.push(send)
-    if (this._closing) this._closeMaybe()
-
-    cb(status < 0 ? new Error('Send failed (status: ' + status + ')') : null)
+  _resolveAndSend (buf, offset, len, port, host, cb) {
+    dns.lookup(host, { family: 4 }, (err, ip) => {
+      if (err) cb(err)
+      else this.send(buf, offset, len, port, ip, cb)
+    })
   }
 
-  _onconnection (port, addr) {
-    const conn = new Connection(this, port, addr, this._nextConnection, this._allowHalfOpen)
-    this.emit('connection', conn)
-    this._nextConnection = b4a.alloc(binding.sizeof_utp_napi_connection_t)
-    return this._nextConnection
+  _resolveAndBind (port, host) {
+    dns.lookup(host, { family: 4 }, (err, ip) => {
+      if (err) this.emit('error', err)
+      else this.bind(port, ip)
+    })
   }
 
-  _onclose () {
-    binding.utp_napi_destroy(this._handle, this._sent.map(toHandle))
-    this._handle = null
-    this.emit('close')
+  _onsend () {
+    if (this._socket.closing) this._closeMaybe()
+  }
+
+  _onconnection (port, ip, handle) {
+    this.emit('connection', new Connection(this, handle, port, ip, this._allowHalfOpen))
   }
 
   static createServer (opts, onconnection) {
@@ -258,30 +173,139 @@ const Socket = module.exports = class Socket extends EventEmitter {
   }
 
   static connect (port, host, opts) {
-    const udp = new Socket(opts)
-    return udp.connect(port, host).on('close', ononeoffclose)
+    const socket = new Socket(opts)
+    return socket
+      .connect(port, host)
+      .on('close', () => socket.close())
   }
 }
 
 Socket.Socket = Socket
 
-function SendRequest () {
-  this._handle = b4a.alloc(binding.sizeof_utp_napi_send_request_t)
-  this._buffer = null
-  this._callback = null
-  this._index = null
-}
+class Connection extends Duplex {
+  constructor (socket, connection, port, address, halfOpen) {
+    super()
 
-function noop () {}
+    this.remoteAddress = address
+    this.remoteFamily = 'IPv4'
+    this.remotePort = port
 
-function isIP (ip) {
-  return IPv4Pattern.test(ip)
-}
+    this._index = -1
+    this._socket = socket
+    this._connection = connection || new UTPConnection(this._socket._socket)
+    this._view = new Uint32Array(this._connection._handle.buffer, this._connection._handle.byteOffset, 2)
+    this._timeout = null
+    this._contentSize = 0
+    this._allowOpen = halfOpen ? 2 : 1
 
-function toHandle (obj) {
-  return obj._handle
-}
+    this._connection
+      .on('data', (buffer) => this._onread(buffer))
+      .on('end', () => this._onend())
+      .on('connect', () => this._onconnect())
+  }
 
-function ononeoffclose () {
-  this._utp.close()
+  setTimeout (ms, ontimeout) {
+    if (ontimeout) this.once('timeout', ontimeout)
+    if (this._timeout) this._timeout.destroy()
+    this._timeout = timeout(ms, this._ontimeout, this)
+  }
+
+  setInteractive (interactive) {
+    this.setPacketSize(interactive ? 0 : 65536)
+  }
+
+  setContentSize (size) {
+    this._view[0] = size < 65536 ? (size >= 0 ? size : 0) : 65536
+    this._contentSize = size
+  }
+
+  setPacketSize (size) {
+    if (size > 65536) size = 65536
+    this._view[0] = size
+    this._contentSize = 0
+  }
+
+  address () {
+    return this.destroyed ? null : this._socket.address()
+  }
+
+  _open (cb) {
+    if (this._connection.connected) cb(null)
+    else {
+      this._connection.once('connect', () => {
+        if (this._timeout) this._timeout.refresh()
+        cb(null)
+      })
+    }
+  }
+
+  _destroy (cb) {
+    this._connection.close(cb)
+  }
+
+  _destroyMaybe () {
+    if (this._allowOpen && !--this._allowOpen) this.destroy()
+  }
+
+  _final (cb) {
+    this._connection.shutdown()
+    this._destroyMaybe()
+    cb(null)
+  }
+
+  _writev (datas, cb) {
+    let bufs = new Array(datas.length)
+    for (var i = 0; i < datas.length; i++) {
+      const data = datas[i]
+      bufs[i] = typeof data === 'string' ? b4a.from(data) : data
+    }
+
+    if (bufs.length > 256) bufs = [b4a.concat(bufs)]
+
+    this._connection.writev(bufs, cb)
+  }
+
+  _connect (port, ip) {
+    this.remotePort = port
+    this.remoteAddress = ip
+
+    this._connection.connect(port, ip, (err) => {
+      if (err) this.emit('error', err)
+    })
+  }
+
+  _resolveAndConnect (port, host) {
+    dns.lookup(host, { family: 4 }, (err, ip) => {
+      if (err) this.destroy(err)
+      else this._connect(port, ip)
+    })
+  }
+
+  _ontimeout () {
+    this.emit('timeout')
+  }
+
+  _onread (buffer) {
+    if (this._timeout) this._timeout.refresh()
+
+    let size = buffer.byteLength
+
+    if (this._contentSize) {
+      if (size > this._contentSize) size = this._contentSize
+      this._contentSize -= size
+      if (this._contentSize < 65536) this._view[0] = this._contentSize
+    }
+
+    this.push(buffer)
+  }
+
+  _onend () {
+    if (this._timeout) this._timeout.destroy()
+    this.push(null)
+    this._destroyMaybe()
+  }
+
+  _onconnect () {
+    this.emit('connect')
+  }
 }
